Lets the Detector class automatically create a GenericDetector class if the detector JSON or dictionary is set up to be used with this class.
